requirements.txt
celery==5.1.0
Django==3.2.4
django-celery-beat==2.2.0
djangorestframework==3.12.4
psycopg2-binary==2.8.6
redis==3.5.3
requests==2.25.1
wikipedia==1.4.0
